Decrypt and process messages all in one transaction.
Giddy up
