+++ app/src/main/java/org/thoughtcrime/securesms/database/PushDatabase.java
@@ -16,6 +16,7 @@ import org.whispersystems.signalservice.api.messages.SignalServiceEnvelope;
 import org.whispersystems.signalservice.api.push.SignalServiceAddress;
 import org.whispersystems.signalservice.internal.util.Util;
 
+import java.io.Closeable;
 import java.io.IOException;
 
 public class PushDatabase extends Database {
@@ -149,7 +150,7 @@ public class PushDatabase extends Database {
     }
   }
 
-  public static class Reader {
+  public static class Reader implements Closeable {
     private final Cursor cursor;
 
     public Reader(Cursor cursor) {
@@ -186,6 +187,7 @@ public class PushDatabase extends Database {
       }
     }
 
+    @Override
     public void close() {
       this.cursor.close();
     }
