+++ app/src/main/java/org/thoughtcrime/securesms/jobs/PushDecryptMessageJob.java
@@ -35,6 +35,11 @@ import org.thoughtcrime.securesms.groups.GroupId;
 import org.thoughtcrime.securesms.jobmanager.Data;
 import org.thoughtcrime.securesms.jobmanager.Job;
 import org.thoughtcrime.securesms.jobmanager.JobManager;
+import org.thoughtcrime.securesms.messages.MessageContentProcessor;
+import org.thoughtcrime.securesms.messages.MessageContentProcessor.ExceptionMetadata;
+import org.thoughtcrime.securesms.messages.MessageContentProcessor.MessageState;
+import org.thoughtcrime.securesms.messages.MessageDecryptionUtil;
+import org.thoughtcrime.securesms.messages.MessageDecryptionUtil.DecryptionResult;
 import org.thoughtcrime.securesms.notifications.NotificationChannels;
 import org.thoughtcrime.securesms.recipients.Recipient;
 import org.thoughtcrime.securesms.transport.RetryLaterException;
@@ -50,8 +55,13 @@ import org.whispersystems.signalservice.internal.push.UnsupportedDataMessageExce
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.LinkedList;
 import java.util.List;
 
+/**
+ * Decrypts an envelope. Enqueues a separate job, {@link PushProcessMessageJob}, to actually insert
+ * the result into our database.
+ */
 public final class PushDecryptMessageJob extends BaseJob {
 
   public static final String KEY   = "PushDecryptJob";
@@ -59,36 +69,36 @@ public final class PushDecryptMessageJob extends BaseJob {
 
   public static final String TAG = Log.tag(PushDecryptMessageJob.class);
 
-  private static final String KEY_MESSAGE_ID     = "message_id";
   private static final String KEY_SMS_MESSAGE_ID = "sms_message_id";
+  private static final String KEY_ENVELOPE       = "envelope";
 
-  private final long messageId;
-  private final long smsMessageId;
+  private final long                  smsMessageId;
+  private final SignalServiceEnvelope envelope;
 
-  public PushDecryptMessageJob(Context context, long pushMessageId) {
-    this(context, pushMessageId, -1);
+  public PushDecryptMessageJob(Context context, @NonNull SignalServiceEnvelope envelope) {
+    this(context, envelope, -1);
   }
 
-  public PushDecryptMessageJob(Context context, long pushMessageId, long smsMessageId) {
+  public PushDecryptMessageJob(Context context, @NonNull SignalServiceEnvelope envelope, long smsMessageId) {
     this(new Parameters.Builder()
                            .setQueue(QUEUE)
                            .setMaxAttempts(Parameters.UNLIMITED)
                            .build(),
-         pushMessageId,
+         envelope,
          smsMessageId);
     setContext(context);
   }
 
-  private PushDecryptMessageJob(@NonNull Parameters parameters, long pushMessageId, long smsMessageId) {
+  private PushDecryptMessageJob(@NonNull Parameters parameters, @NonNull SignalServiceEnvelope envelope, long smsMessageId) {
     super(parameters);
 
-    this.messageId    = pushMessageId;
+    this.envelope     = envelope;
     this.smsMessageId = smsMessageId;
   }
 
   @Override
   public @NonNull Data serialize() {
-    return new Data.Builder().putLong(KEY_MESSAGE_ID, messageId)
+    return new Data.Builder().putBlobAsString(KEY_ENVELOPE, envelope.serialize())
                              .putLong(KEY_SMS_MESSAGE_ID, smsMessageId)
                              .build();
   }
@@ -99,32 +109,30 @@ public final class PushDecryptMessageJob extends BaseJob {
   }
 
   @Override
-  public void onRun() throws NoSuchMessageException, RetryLaterException {
+  public void onRun() throws RetryLaterException {
     if (needsMigration()) {
       Log.w(TAG, "Migration is still needed.");
       postMigrationNotification();
       throw new RetryLaterException();
     }
 
-    PushDatabase          database   = DatabaseFactory.getPushDatabase(context);
-    SignalServiceEnvelope envelope   = database.get(messageId);
-    JobManager            jobManager = ApplicationDependencies.getJobManager();
+    List<Job> jobs = new LinkedList<>();
 
-    try {
-      List<Job> jobs;
+    try (DatabaseSessionLock.Lock unused = DatabaseSessionLock.INSTANCE.acquire()) {
+      DecryptionResult result = MessageDecryptionUtil.decrypt(context, envelope);
 
-      try (DatabaseSessionLock.Lock unused = DatabaseSessionLock.INSTANCE.acquire()) {
-        jobs = handleMessage(envelope);
+      if (result.getContent() != null) {
+        jobs.add(new PushProcessMessageJob(result.getContent(), smsMessageId, envelope.getTimestamp()));
+      } else if (result.getException() != null && result.getState() != MessageState.NOOP) {
+        jobs.add(new PushProcessMessageJob(result.getState(), result.getException(), smsMessageId, envelope.getTimestamp()));
       }
 
-      for (Job job: jobs) {
-        jobManager.add(job);
-      }
-    } catch (NoSenderException e) {
-      Log.w(TAG, "Invalid message, but no sender info!");
+      jobs.addAll(result.getJobs());
     }
 
-    database.delete(messageId);
+    for (Job job: jobs) {
+      ApplicationDependencies.getJobManager().add(job);
+    }
   }
 
   @Override
@@ -141,7 +149,6 @@ public final class PushDecryptMessageJob extends BaseJob {
   }
 
   private void postMigrationNotification() {
-    // TODO [greyson] Navigation
     NotificationManagerCompat.from(context).notify(494949,
                                                    new NotificationCompat.Builder(context, NotificationChannels.getMessagesChannel(context))
                                                                          .setSmallIcon(R.drawable.ic_notification)
@@ -155,110 +162,12 @@ public final class PushDecryptMessageJob extends BaseJob {
 
   }
 
-  private @NonNull List<Job> handleMessage(@NonNull SignalServiceEnvelope envelope) throws NoSenderException {
-    Log.i(TAG, "Processing message ID " + envelope.getTimestamp());
-    try {
-      SignalProtocolStore  axolotlStore = new SignalProtocolStoreImpl(context);
-      SignalServiceAddress localAddress = new SignalServiceAddress(Optional.of(TextSecurePreferences.getLocalUuid(context)), Optional.of(TextSecurePreferences.getLocalNumber(context)));
-      SignalServiceCipher  cipher       = new SignalServiceCipher(localAddress, axolotlStore, DatabaseSessionLock.INSTANCE, UnidentifiedAccessUtil.getCertificateValidator());
-
-      SignalServiceContent content = cipher.decrypt(envelope);
-
-      List<Job> jobs = new ArrayList<>(2);
-
-      if (content != null) {
-        jobs.add(new PushProcessMessageJob(content, messageId, smsMessageId, envelope.getTimestamp()));
-      }
-
-      if (envelope.isPreKeySignalMessage()) {
-        jobs.add(new RefreshPreKeysJob());
-      }
-
-      return jobs;
-
-    } catch (ProtocolInvalidVersionException e) {
-      Log.w(TAG, String.valueOf(envelope.getTimestamp()), e);
-      return Collections.singletonList(new PushProcessMessageJob(PushProcessMessageJob.MessageState.INVALID_VERSION,
-                                                                 toExceptionMetadata(e),
-                                                                 messageId,
-                                                                 smsMessageId,
-                                                                 envelope.getTimestamp()));
-
-    } catch (ProtocolInvalidMessageException | ProtocolInvalidKeyIdException | ProtocolInvalidKeyException | ProtocolUntrustedIdentityException | ProtocolNoSessionException e) {
-      Log.w(TAG, String.valueOf(envelope.getTimestamp()), e);
-      return Collections.singletonList(new AutomaticSessionResetJob(Recipient.external(context, e.getSender()).getId(),
-                                                                    e.getSenderDevice(),
-                                                                    envelope.getTimestamp()));
-    } catch (ProtocolLegacyMessageException e) {
-      Log.w(TAG, String.valueOf(envelope.getTimestamp()), e);
-      return Collections.singletonList(new PushProcessMessageJob(PushProcessMessageJob.MessageState.LEGACY_MESSAGE,
-                                                                 toExceptionMetadata(e),
-                                                                 messageId,
-                                                                 smsMessageId,
-                                                                 envelope.getTimestamp()));
-
-    } catch (ProtocolDuplicateMessageException e) {
-      Log.w(TAG, String.valueOf(envelope.getTimestamp()), e);
-      return Collections.singletonList(new PushProcessMessageJob(PushProcessMessageJob.MessageState.DUPLICATE_MESSAGE,
-                                                                 toExceptionMetadata(e),
-                                                                 messageId,
-                                                                 smsMessageId,
-                                                                 envelope.getTimestamp()));
-
-    } catch (InvalidMetadataVersionException | InvalidMetadataMessageException e) {
-      Log.w(TAG, String.valueOf(envelope.getTimestamp()), e);
-      return Collections.emptyList();
-
-    } catch (SelfSendException e) {
-      Log.i(TAG, "Dropping UD message from self.");
-      return Collections.emptyList();
-
-    } catch (UnsupportedDataMessageException e) {
-      Log.w(TAG, String.valueOf(envelope.getTimestamp()), e);
-      return Collections.singletonList(new PushProcessMessageJob(PushProcessMessageJob.MessageState.UNSUPPORTED_DATA_MESSAGE,
-                                       toExceptionMetadata(e),
-                                       messageId,
-                                       smsMessageId,
-                                       envelope.getTimestamp()));
-    }
-  }
-
-  private static PushProcessMessageJob.ExceptionMetadata toExceptionMetadata(@NonNull UnsupportedDataMessageException e)
-      throws NoSenderException
-  {
-    String sender = e.getSender();
-
-    if (sender == null) throw new NoSenderException();
-
-    GroupId groupId = null;
-
-    if (e.getGroup().isPresent()) {
-      try {
-        groupId = GroupUtil.idFromGroupContext(e.getGroup().get());
-      } catch (BadGroupIdException ex) {
-        Log.w(TAG, "Bad group id found in unsupported data message", ex);
-      }
-    }
-
-    return new PushProcessMessageJob.ExceptionMetadata(sender,
-                                                       e.getSenderDevice(),
-                                                       groupId);
-  }
-
-  private static PushProcessMessageJob.ExceptionMetadata toExceptionMetadata(@NonNull ProtocolException e) throws NoSenderException {
-    String sender = e.getSender();
-
-    if (sender == null) throw new NoSenderException();
-
-    return new PushProcessMessageJob.ExceptionMetadata(sender, e.getSenderDevice());
-  }
-
   public static final class Factory implements Job.Factory<PushDecryptMessageJob> {
     @Override
     public @NonNull PushDecryptMessageJob create(@NonNull Parameters parameters, @NonNull Data data) {
-      return new PushDecryptMessageJob(parameters, data.getLong(KEY_MESSAGE_ID), data.getLong(KEY_SMS_MESSAGE_ID));
+      return new PushDecryptMessageJob(parameters,
+                                       SignalServiceEnvelope.deserialize(data.getStringAsBlob(KEY_ENVELOPE)),
+                                       data.getLong(KEY_SMS_MESSAGE_ID));
     }
   }
-
-  private static class NoSenderException extends Exception {}
 }
