+++ app/src/main/java/org/thoughtcrime/securesms/util/GroupUtil.java
@@ -21,6 +21,7 @@ import org.thoughtcrime.securesms.mms.OutgoingGroupUpdateMessage;
 import org.thoughtcrime.securesms.recipients.Recipient;
 import org.thoughtcrime.securesms.recipients.RecipientId;
 import org.whispersystems.libsignal.util.guava.Optional;
+import org.whispersystems.signalservice.api.messages.SignalServiceContent;
 import org.whispersystems.signalservice.api.messages.SignalServiceDataMessage;
 import org.whispersystems.signalservice.api.messages.SignalServiceGroup;
 import org.whispersystems.signalservice.api.messages.SignalServiceGroupContext;
@@ -37,6 +38,24 @@ public final class GroupUtil {
   }
 
   private static final String TAG = Log.tag(GroupUtil.class);
+
+  /**
+   * @return The group context present on the content if one exists, otherwise null.
+   */
+  public static @Nullable SignalServiceGroupContext getGroupContextIfPresent(@Nullable SignalServiceContent content) {
+    if (content == null) {
+      return null;
+    } else if (content.getDataMessage().isPresent() && content.getDataMessage().get().getGroupContext().isPresent()) {
+      return content.getDataMessage().get().getGroupContext().get();
+    } else if (content.getSyncMessage().isPresent()                 &&
+               content.getSyncMessage().get().getSent().isPresent() &&
+               content.getSyncMessage().get().getSent().get().getMessage().getGroupContext().isPresent())
+    {
+      return content.getSyncMessage().get().getSent().get().getMessage().getGroupContext().get();
+    } else {
+      return null;
+    }
+  }
 
   /**
    * Result may be a v1 or v2 GroupId.
