+++ libsignal/service/src/main/java/org/whispersystems/signalservice/api/messages/SignalServiceEnvelope.java
@@ -7,6 +7,7 @@
 package org.whispersystems.signalservice.api.messages;
 
 import com.google.protobuf.ByteString;
+import com.google.protobuf.InvalidProtocolBufferException;
 
 import org.whispersystems.libsignal.InvalidVersionException;
 import org.whispersystems.libsignal.logging.Log;
@@ -14,6 +15,7 @@ import org.whispersystems.libsignal.util.guava.Optional;
 import org.whispersystems.signalservice.api.push.SignalServiceAddress;
 import org.whispersystems.signalservice.api.util.UuidUtil;
 import org.whispersystems.signalservice.internal.push.SignalServiceProtos.Envelope;
+import org.whispersystems.signalservice.internal.serialize.protos.SignalServiceEnvelopeProto;
 import org.whispersystems.signalservice.internal.util.Hex;
 import org.whispersystems.util.Base64;
 
@@ -257,5 +259,55 @@ public class SignalServiceEnvelope {
 
   public boolean isUnidentifiedSender() {
     return envelope.getType().getNumber() == Envelope.Type.UNIDENTIFIED_SENDER_VALUE;
+  }
+
+  public byte[] serialize() {
+    SignalServiceEnvelopeProto.Builder builder = SignalServiceEnvelopeProto.newBuilder()
+                                                                           .setType(getType())
+                                                                           .setDeviceId(getSourceDevice())
+                                                                           .setTimestamp(getTimestamp())
+                                                                           .setServerReceivedTimestamp(getServerReceivedTimestamp())
+                                                                           .setServerDeliveredTimestamp(getServerDeliveredTimestamp());
+
+    if (getSourceUuid().isPresent()) {
+      builder.setSourceUuid(getSourceUuid().get());
+    }
+
+    if (getSourceE164().isPresent()) {
+      builder.setSourceE164(getSourceE164().get());
+    }
+
+    if (hasLegacyMessage()) {
+      builder.setLegacyMessage(ByteString.copyFrom(getLegacyMessage()));
+    }
+
+    if (hasContent()) {
+      builder.setContent(ByteString.copyFrom(getContent()));
+    }
+
+    if (hasUuid()) {
+      builder.setServerGuid(getUuid());
+    }
+
+    return builder.build().toByteArray();
+  }
+
+  public static SignalServiceEnvelope deserialize(byte[] serialized) {
+    SignalServiceEnvelopeProto proto = null;
+    try {
+      proto = SignalServiceEnvelopeProto.parseFrom(serialized);
+    } catch (InvalidProtocolBufferException e) {
+      e.printStackTrace();
+    }
+
+    return new SignalServiceEnvelope(proto.getType(),
+                                     SignalServiceAddress.fromRaw(proto.getSourceUuid(), proto.getSourceE164()),
+                                     proto.getDeviceId(),
+                                     proto.getTimestamp(),
+                                     proto.hasLegacyMessage() ? proto.getLegacyMessage().toByteArray() : null,
+                                     proto.hasContent() ? proto.getContent().toByteArray() : null,
+                                     proto.getServerReceivedTimestamp(),
+                                     proto.getServerDeliveredTimestamp(),
+                                     proto.getServerGuid());
   }
 }
